Database: passing dependencies through constructor
